+++ src/main/java/demo/mapper/user/UserMapper.java
@@ -11,7 +11,7 @@ public interface UserMapper {
 
     Integer addUser(UserProfileDto userProfileDto);
     Integer updateUser(UserProfileDto userProfileDto);
-    UserProfileDto getUserDetail(Integer id);
+    UserProfileDto getUserDetail(Long id);
     Integer deleteUser(UserProfileDto userProfileDto);
 }
